authOptions moved another file for building the project with an healthy way
