+++ app/api/auth/[...nextauth]/route.ts
@@ -1,35 +1,5 @@
 import NextAuth from "next-auth";
-import Auth0Provider from "next-auth/providers/auth0";
-import type { NextAuthOptions } from "next-auth";
-
-export const authOptions: NextAuthOptions = {
-    providers: [
-        Auth0Provider({
-            clientId: process.env.AUTH0_CLIENT_ID!,
-            clientSecret: process.env.AUTH0_CLIENT_SECRET!,
-            issuer: process.env.AUTH0_ISSUER,
-        }),
-    ],
-    session: {
-        strategy: "jwt",
-    },
-    secret: process.env.NEXTAUTH_SECRET,
-    callbacks: {
-        async jwt({ token, user, account }) {
-            if (account && user) {
-                token.accessToken = account.access_token;
-                token.id = user.id;
-                token.email = user.email;
-            }
-            return token;
-        },
-        async session({ session, token }) {
-            session.user.id = token.id as string;
-            session.user.email = token.email as string;
-            return session;
-        }
-    }
-};
+import { authOptions } from "@/server/auth";
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
