+++ server/auth.ts
@@ -1,0 +1,31 @@
+import type { NextAuthOptions } from "next-auth";
+import Auth0Provider from "next-auth/providers/auth0";
+
+export const authOptions: NextAuthOptions = {
+    providers: [
+        Auth0Provider({
+            clientId: process.env.AUTH0_CLIENT_ID!,
+            clientSecret: process.env.AUTH0_CLIENT_SECRET!,
+            issuer: process.env.AUTH0_ISSUER,
+        }),
+    ],
+    session: {
+        strategy: "jwt",
+    },
+    secret: process.env.NEXTAUTH_SECRET,
+    callbacks: {
+        async jwt({ token, user, account }) {
+            if (account && user) {
+                token.accessToken = account.access_token;
+                token.id = user.id;
+                token.email = user.email;
+            }
+            return token;
+        },
+        async session({ session, token }) {
+            session.user.id = token.id as string;
+            session.user.email = token.email as string;
+            return session;
+        }
+    }
+};
